+++ src/data/pokemon/tmhm_learnsets.h
@@ -285,6 +285,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_WEEDLE] = { .learnset = {
@@ -542,6 +543,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_RAICHU] = { .learnset = {
@@ -569,6 +572,9 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_SANDSHREW] = { .learnset = {
@@ -856,6 +862,8 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_CLEFABLE] = { .learnset = {
@@ -894,6 +902,8 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_VULPIX] = { .learnset = {
@@ -973,6 +983,7 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_WIGGLYTUFF] = { .learnset = {
@@ -1009,6 +1020,7 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_ZUBAT] = { .learnset = {
@@ -1078,6 +1090,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GLOOM] = { .learnset = {
@@ -1098,6 +1111,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_VILEPLUME] = { .learnset = {
@@ -1119,6 +1133,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PARAS] = { .learnset = {
@@ -1604,6 +1619,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_KADABRA] = { .learnset = {
@@ -1634,6 +1650,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_ALAKAZAM] = { .learnset = {
@@ -1665,6 +1682,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MACHOP] = { .learnset = {
@@ -1833,6 +1851,7 @@ const union {
         .SURF = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_TENTACRUEL] = { .learnset = {
@@ -1859,6 +1878,7 @@ const union {
         .SURF = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GEODUDE] = { .learnset = {
@@ -2315,6 +2335,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_HAUNTER] = { .learnset = {
@@ -2340,6 +2361,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GENGAR] = { .learnset = {
@@ -2371,6 +2393,7 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_ONIX] = { .learnset = {
@@ -2424,6 +2447,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_HYPNO] = { .learnset = {
@@ -2454,6 +2479,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_KRABBY] = { .learnset = {
@@ -2908,6 +2935,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_TANGELA] = { .learnset = {
@@ -3084,6 +3113,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_STARMIE] = { .learnset = {
@@ -3112,6 +3142,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MR_MIME] = { .learnset = {
@@ -3146,6 +3177,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SCYTHER] = { .learnset = {
@@ -3198,6 +3230,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_ELECTABUZZ] = { .learnset = {
@@ -3374,6 +3407,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE
     } },
 
     [SPECIES_DITTO] = { .learnset = {
@@ -3956,6 +3990,9 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_CHIKORITA] = { .learnset = {
@@ -4450,6 +4487,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_LANTURN] = { .learnset = {
@@ -4476,6 +4514,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PICHU] = { .learnset = {
@@ -4496,6 +4535,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_CLEFFA] = { .learnset = {
@@ -4524,6 +4564,9 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_IGGLYBUFF] = { .learnset = {
@@ -4551,6 +4594,9 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_TOGEPI] = { .learnset = {
@@ -4578,6 +4624,8 @@ const union {
         .ATTRACT = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_TOGETIC] = { .learnset = {
@@ -4611,6 +4659,8 @@ const union {
         .FLY = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_NATU] = { .learnset = {
@@ -4638,6 +4688,7 @@ const union {
         .STEEL_WING = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_XATU] = { .learnset = {
@@ -4667,6 +4718,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .FLY = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MAREEP] = { .learnset = {
@@ -4687,6 +4739,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_FLAAFFY] = { .learnset = {
@@ -4711,6 +4764,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_AMPHAROS] = { .learnset = {
@@ -4736,6 +4790,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BELLOSSOM] = { .learnset = {
@@ -4758,6 +4813,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MARILL] = { .learnset = {
@@ -4785,6 +4841,8 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_AZUMARILL] = { .learnset = {
@@ -4813,6 +4871,8 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_SUDOWOODO] = { .learnset = {
@@ -4887,6 +4947,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SKIPLOOM] = { .learnset = {
@@ -4906,6 +4967,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_JUMPLUFF] = { .learnset = {
@@ -4926,6 +4988,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_AIPOM] = { .learnset = {
@@ -5112,6 +5175,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_UMBREON] = { .learnset = {
@@ -5431,6 +5496,7 @@ const union {
         .OVERHEAT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GRANBULL] = { .learnset = {
@@ -5470,6 +5536,7 @@ const union {
         .OVERHEAT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_QWILFISH] = { .learnset = {
@@ -6161,6 +6228,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_ELEKID] = { .learnset = {
@@ -6282,6 +6350,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_RAIKOU] = { .learnset = {
@@ -7081,6 +7151,7 @@ const union {
         .THIEF = TRUE,
         .SURF = TRUE,
         .FLASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_LOMBRE] = { .learnset = {
@@ -7111,6 +7182,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_LUDICOLO] = { .learnset = {
@@ -7143,6 +7215,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_SEEDOT] = { .learnset = {
@@ -7737,6 +7810,7 @@ const union {
         .ATTRACT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_TORKOAL] = { .learnset = {
@@ -7790,6 +7864,7 @@ const union {
         .CUT = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BARBOACH] = { .learnset = {
@@ -8492,6 +8567,7 @@ const union {
         .ATTRACT = TRUE,
         .SURF = TRUE,
         .WATERFALL = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_SPOINK] = { .learnset = {
@@ -8520,6 +8596,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GRUMPIG] = { .learnset = {
@@ -8551,6 +8628,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PLUSLE] = { .learnset = {
@@ -8620,6 +8698,7 @@ const union {
         .ATTRACT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_MEDITITE] = { .learnset = {
@@ -8697,6 +8776,8 @@ const union {
         .THIEF = TRUE,
         .STEEL_WING = TRUE,
         .FLY = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_ALTARIA] = { .learnset = {
@@ -8727,6 +8808,8 @@ const union {
         .STEEL_WING = TRUE,
         .FLY = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_WYNAUT] = { .learnset = {
@@ -8809,6 +8892,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SLAKOTH] = { .learnset = {
@@ -9215,6 +9299,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BANETTE] = { .learnset = {
@@ -9243,6 +9328,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SEVIPER] = { .learnset = {
@@ -9485,6 +9571,7 @@ const union {
         .ATTRACT = TRUE,
         .THIEF = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_ILLUMISE] = { .learnset = {
@@ -9513,6 +9600,9 @@ const union {
         .ATTRACT = TRUE,
         .THIEF = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_LILEEP] = { .learnset = {
@@ -9633,6 +9723,9 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE= TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_KIRLIA] = { .learnset = {
@@ -9662,6 +9755,9 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+		.DISARMING_VOICE= TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_GARDEVOIR] = { .learnset = {
@@ -9692,6 +9788,9 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+		.DISARMING_VOICE= TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_BAGON] = { .learnset = {
@@ -10127,6 +10226,7 @@ const union {
         .REST = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_DEOXYS] = { .learnset = {
@@ -10193,6 +10293,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },    
 
     [SPECIES_AMBIPOM] = { .learnset = {
@@ -10284,6 +10386,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
 
     [SPECIES_BONSLY] = { .learnset = {
@@ -10328,6 +10431,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .CUT = TRUE,
+		.DAZZLING_GLEAM = TRUE
     } }, 
 
     [SPECIES_CHINGLING] = { .learnset = {
@@ -10606,6 +10710,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } }, 
         
     [SPECIES_GLACEON] = { .learnset = {
@@ -11161,7 +11267,8 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .CUT = TRUE,
-        .FLASH = TRUE
+        .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
         
     [SPECIES_SYLVEON] = { .learnset = {
@@ -11186,7 +11293,10 @@ const union {
         .ATTRACT = TRUE,
         .THIEF = TRUE,
         .CUT = TRUE,
-        .STRENGTH = TRUE
+        .STRENGTH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } }, 
         
     [SPECIES_TANGROWTH] = { .learnset = {
@@ -11249,7 +11359,8 @@ const union {
         .STEEL_WING = TRUE,
         .FLY = TRUE,
         .FLASH = TRUE,
-        .ROCK_SMASH = TRUE
+        .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
         
     [SPECIES_WEAVILE] = { .learnset = {
@@ -11340,7 +11451,8 @@ const union {
         .THIEF = TRUE,
         .OVERHEAT = TRUE,
         .STRENGTH = TRUE,
-        .ROCK_SMASH = TRUE
+        .ROCK_SMASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },                                  
 
     [SPECIES_FARIGIRAF] = { .learnset = {
@@ -11546,6 +11658,7 @@ const union {
         .FLASH = TRUE,
 		.BULLET_SEED = TRUE,
 		.DIG = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 	
 	[SPECIES_STUFFUL] = { .learnset = {
@@ -11932,6 +12045,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_WEEDLE] = { .learnset = {
@@ -12189,6 +12303,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_RAICHU] = { .learnset = {
@@ -12216,6 +12332,9 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_SANDSHREW] = { .learnset = {
@@ -12503,6 +12622,9 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_CLEFABLE] = { .learnset = {
@@ -12541,6 +12663,9 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_VULPIX] = { .learnset = {
@@ -12619,6 +12744,8 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_WIGGLYTUFF] = { .learnset = {
@@ -12655,6 +12782,8 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_ZUBAT] = { .learnset = {
@@ -12724,6 +12853,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GLOOM] = { .learnset = {
@@ -12744,6 +12874,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_VILEPLUME] = { .learnset = {
@@ -12765,6 +12896,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PARAS] = { .learnset = {
@@ -13249,6 +13381,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_KADABRA] = { .learnset = {
@@ -13279,6 +13412,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_ALAKAZAM] = { .learnset = {
@@ -13310,6 +13444,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MACHOP] = { .learnset = {
@@ -13478,6 +13613,7 @@ const union {
         .SURF = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_TENTACRUEL] = { .learnset = {
@@ -13504,6 +13640,7 @@ const union {
         .SURF = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GEODUDE] = { .learnset = {
@@ -13960,6 +14097,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_HAUNTER] = { .learnset = {
@@ -13985,6 +14123,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GENGAR] = { .learnset = {
@@ -14016,6 +14155,7 @@ const union {
         .SNATCH = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_ONIX] = { .learnset = {
@@ -14069,6 +14209,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_HYPNO] = { .learnset = {
@@ -14099,6 +14241,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_KRABBY] = { .learnset = {
@@ -14551,6 +14695,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_TANGELA] = { .learnset = {
@@ -14727,6 +14873,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_STARMIE] = { .learnset = {
@@ -14755,6 +14902,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MR_MIME] = { .learnset = {
@@ -14789,6 +14937,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SCYTHER] = { .learnset = {
@@ -14841,6 +14990,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_ELECTABUZZ] = { .learnset = {
@@ -15017,6 +15167,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_DITTO] = { .learnset = {
@@ -15599,6 +15750,9 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_CHIKORITA] = { .learnset = {
@@ -15622,6 +15776,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BAYLEEF] = { .learnset = {
@@ -15647,6 +15802,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MEGANIUM] = { .learnset = {
@@ -15674,6 +15830,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_CYNDAQUIL] = { .learnset = {
@@ -16091,6 +16248,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_LANTURN] = { .learnset = {
@@ -16117,6 +16275,7 @@ const union {
         .FLASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PICHU] = { .learnset = {
@@ -16137,6 +16296,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_CLEFFA] = { .learnset = {
@@ -16165,6 +16325,9 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_IGGLYBUFF] = { .learnset = {
@@ -16192,6 +16355,9 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_TOGEPI] = { .learnset = {
@@ -16219,6 +16385,8 @@ const union {
         .ATTRACT = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_TOGETIC] = { .learnset = {
@@ -16252,6 +16420,8 @@ const union {
         .FLY = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_NATU] = { .learnset = {
@@ -16279,6 +16449,7 @@ const union {
         .STEEL_WING = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_XATU] = { .learnset = {
@@ -16308,6 +16479,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .FLY = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MAREEP] = { .learnset = {
@@ -16328,6 +16500,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_FLAAFFY] = { .learnset = {
@@ -16352,6 +16525,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_AMPHAROS] = { .learnset = {
@@ -16377,6 +16551,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BELLOSSOM] = { .learnset = {
@@ -16399,6 +16574,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_MARILL] = { .learnset = {
@@ -16426,6 +16602,8 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_AZUMARILL] = { .learnset = {
@@ -16454,6 +16632,8 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_SUDOWOODO] = { .learnset = {
@@ -16528,6 +16708,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SKIPLOOM] = { .learnset = {
@@ -16547,6 +16728,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_JUMPLUFF] = { .learnset = {
@@ -16567,6 +16749,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_AIPOM] = { .learnset = {
@@ -16644,6 +16827,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_YANMA] = { .learnset = {
@@ -16753,6 +16937,8 @@ const union {
         .SKILL_SWAP = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_UMBREON] = { .learnset = {
@@ -16867,6 +17053,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_UNOWN] = { .learnset = {
@@ -16903,6 +17090,7 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PINECO] = { .learnset = {
@@ -17072,6 +17260,7 @@ const union {
         .OVERHEAT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GRANBULL] = { .learnset = {
@@ -17111,6 +17300,7 @@ const union {
         .OVERHEAT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_QWILFISH] = { .learnset = {
@@ -17802,6 +17992,7 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_ELEKID] = { .learnset = {
@@ -17923,6 +18114,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_RAIKOU] = { .learnset = {
@@ -18205,6 +18398,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_OLD_UNOWN_B] = { .learnset = {
@@ -18720,6 +18914,7 @@ const union {
         .THIEF = TRUE,
         .SURF = TRUE,
         .FLASH = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_LOMBRE] = { .learnset = {
@@ -18750,6 +18945,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_LUDICOLO] = { .learnset = {
@@ -18782,6 +18978,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_SEEDOT] = { .learnset = {
@@ -19373,6 +19570,7 @@ const union {
         .ATTRACT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_TORKOAL] = { .learnset = {
@@ -19426,6 +19624,7 @@ const union {
         .CUT = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BARBOACH] = { .learnset = {
@@ -20126,6 +20325,7 @@ const union {
         .ATTRACT = TRUE,
         .SURF = TRUE,
         .WATERFALL = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_SPOINK] = { .learnset = {
@@ -20154,6 +20354,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_GRUMPIG] = { .learnset = {
@@ -20184,6 +20385,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_PLUSLE] = { .learnset = {
@@ -20253,6 +20455,7 @@ const union {
         .ATTRACT = TRUE,
         .STRENGTH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_MEDITITE] = { .learnset = {
@@ -20330,6 +20533,8 @@ const union {
         .THIEF = TRUE,
         .STEEL_WING = TRUE,
         .FLY = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_ALTARIA] = { .learnset = {
@@ -20360,6 +20565,8 @@ const union {
         .STEEL_WING = TRUE,
         .FLY = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } },
 
     [SPECIES_WYNAUT] = { .learnset = {
@@ -20442,6 +20649,7 @@ const union {
         .ATTRACT = TRUE,
         .CUT = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SLAKOTH] = { .learnset = {
@@ -20848,6 +21056,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_BANETTE] = { .learnset = {
@@ -20876,6 +21085,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_SEVIPER] = { .learnset = {
@@ -21117,6 +21327,7 @@ const union {
         .ATTRACT = TRUE,
         .THIEF = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_ILLUMISE] = { .learnset = {
@@ -21145,6 +21356,9 @@ const union {
         .ATTRACT = TRUE,
         .THIEF = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_LILEEP] = { .learnset = {
@@ -21265,6 +21479,9 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_KIRLIA] = { .learnset = {
@@ -21294,6 +21511,9 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_GARDEVOIR] = { .learnset = {
@@ -21324,6 +21544,9 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } },
 
     [SPECIES_BAGON] = { .learnset = {
@@ -21759,6 +21982,7 @@ const union {
         .REST = TRUE,
         .SKILL_SWAP = TRUE,
         .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_DEOXYS] = { .learnset = {
@@ -21825,6 +22049,7 @@ const union {
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
         .FLASH = TRUE,
+		.DAZZLING_GLEAM = TRUE,
     } },
 
     [SPECIES_AMBIPOM] = { .learnset = {
@@ -21916,6 +22141,7 @@ const union {
         .ROCK_SMASH = TRUE,
         .WATERFALL = TRUE,
         .DIVE = TRUE,
+		.DAZZLING_GLEAM = TRUE
     } }, 
 
     [SPECIES_BONSLY] = { .learnset = {
@@ -21960,6 +22186,7 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .CUT = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
 
     [SPECIES_CHINGLING] = { .learnset = {
@@ -22238,6 +22465,8 @@ const union {
         .STRENGTH = TRUE,
         .FLASH = TRUE,
         .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
     } }, 
         
     [SPECIES_GLACEON] = { .learnset = {
@@ -22571,7 +22800,8 @@ const union {
         .THIEF = TRUE,
         .SKILL_SWAP = TRUE,
         .SNATCH = TRUE,
-        .FLASH = TRUE
+        .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
         
     [SPECIES_MUNCHLAX] = { .learnset = {
@@ -22793,7 +23023,8 @@ const union {
         .REST = TRUE,
         .ATTRACT = TRUE,
         .CUT = TRUE,
-        .FLASH = TRUE
+        .FLASH = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
         
     [SPECIES_SYLVEON] = { .learnset = {
@@ -22818,7 +23049,10 @@ const union {
         .ATTRACT = TRUE,
         .THIEF = TRUE,
         .CUT = TRUE,
-        .STRENGTH = TRUE
+        .STRENGTH = TRUE,
+        .DAZZLING_GLEAM = TRUE,
+        .DISARMING_VOICE = TRUE,
+        .DRAINING_KISS = TRUE,
     } }, 
         
     [SPECIES_TANGROWTH] = { .learnset = {
@@ -22881,7 +23115,8 @@ const union {
         .STEEL_WING = TRUE,
         .FLY = TRUE,
         .FLASH = TRUE,
-        .ROCK_SMASH = TRUE
+        .ROCK_SMASH = TRUE,
+        .DAZZLING_GLEAM = TRUE
     } }, 
         
     [SPECIES_WEAVILE] = { .learnset = {